+++ socketsecurity/__init__.py
@@ -1,3 +1,3 @@
 __author__ = 'socket.dev'
-__version__ = '2.4.9'
+__version__ = '2.4.12'
 USER_AGENT = f'SocketPythonCLI/{__version__}'